Fix build: cast NextAnimation to uint after widening row values to int

+++ WadTool/Forms/FormStateChangesEditor.cs
@@ -252,7 +252,7 @@ namespace WadTool
                 if (anim != null)
                     cell.ToolTipText = anim.WadAnimation.Name;
                 else
-                    cell.ToolTipText = TrCatalog.GetAnimationName(_editor.Tool.DestinationWad.GameVersion, _editor.Moveable.Id.TypeId, item.NextAnimation);
+                    cell.ToolTipText = TrCatalog.GetAnimationName(_editor.Tool.DestinationWad.GameVersion, _editor.Moveable.Id.TypeId, (uint)item.NextAnimation);
             }
         }
